requirements.txt
discord
pillow